+++ src/shubhamG/Assignment31to40/Assignment33NumberWithItsFrequency.java
@@ -1,4 +1,4 @@
-package shubhamG.Assignment21to30;
+package shubhamG.Assignment31to40;
 
 import java.util.Arrays;
